refactor: separate GrantsMain component to add a dedicated metadata in server side
